Resolve circular import between shields and question_validity.
Move append_turn_to_conversation into utils.conversations so capabilities
no longer depend on utils.shields at import time.
